Remove warnings from MediaBrowser.Controller
