+++ MediaBrowser.Controller/Channels/IDisableMediaSourceDisplay.cs
@@ -1,7 +1,13 @@
-﻿#pragma warning disable CS1591
+﻿#pragma warning disable CA1040 // Avoid empty interfaces
 
 namespace MediaBrowser.Controller.Channels
 {
+    /// <summary>
+    /// Disable media source display.
+    /// </summary>
+    /// <remarks>
+    /// <see cref="Channel"/> can inherit this interface to disable being displayed.
+    /// </remarks>
     public interface IDisableMediaSourceDisplay
     {
     }
