+++ MediaBrowser.Controller/Channels/ISupportsMediaProbe.cs
@@ -1,7 +1,10 @@
-﻿#pragma warning disable CS1591
+﻿#pragma warning disable CA1040 // Avoid empty interfaces
 
 namespace MediaBrowser.Controller.Channels
 {
+    /// <summary>
+    /// Channel supports media probe.
+    /// </summary>
     public interface ISupportsMediaProbe
     {
     }
