+++ MediaBrowser.Controller/Entities/IHasScreenshots.cs
@@ -1,7 +1,9 @@
+#pragma warning disable CA1040 // Avoid empty interfaces
+
 namespace MediaBrowser.Controller.Entities
 {
     /// <summary>
-    /// Interface IHasScreenshots.
+    /// The item has screenshots.
     /// </summary>
     public interface IHasScreenshots
     {
