+++ MediaBrowser.Controller/Entities/ISupportsBoxSetGrouping.cs
@@ -1,3 +1,5 @@
+#pragma warning disable CA1040 // Avoid empty interfaces
+
 namespace MediaBrowser.Controller.Entities
 {
     /// <summary>
