+++ MediaBrowser.Controller/Net/BasePeriodicWebSocketListener.cs
@@ -270,13 +270,4 @@ namespace MediaBrowser.Controller.Net
             GC.SuppressFinalize(this);
         }
     }
-
-    public class WebSocketListenerState
-    {
-        public DateTime DateLastSendUtc { get; set; }
-
-        public long InitialDelayMs { get; set; }
-
-        public long IntervalMs { get; set; }
-    }
 }
