+++ MediaBrowser.Controller/Persistence/IUserDataRepository.cs
@@ -40,17 +40,16 @@ namespace MediaBrowser.Controller.Persistence
         /// <summary>
         /// Return all user data associated with the given user.
         /// </summary>
-        /// <param name="userId"></param>
-        /// <returns></returns>
+        /// <param name="userId">The user id.</param>
+        /// <returns>The list of user item data.</returns>
         List<UserItemData> GetAllUserData(long userId);
 
         /// <summary>
         /// Save all user data associated with the given user.
         /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="userData"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
+        /// <param name="userId">The user id.</param>
+        /// <param name="userData">The user item data.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
         void SaveAllUserData(long userId, UserItemData[] userData, CancellationToken cancellationToken);
     }
 }
