+++ MediaBrowser.Controller/Providers/IForcedProvider.cs
@@ -1,3 +1,5 @@
+#pragma warning disable CA1040 // Avoid empty interfaces
+
 namespace MediaBrowser.Controller.Providers
 {
     /// <summary>
